+++ celeryconfig.py
@@ -36,7 +36,7 @@ if os.environ.get("OQ_ENGINE_USE_SRCDIR"):
     sys.modules['openquake'].__dict__["__path__"].insert(
         0, os.path.join(os.path.dirname(__file__), "openquake"))
 
-from openquake.engine.utils import config
+from openquake.engine import config
 
 config.abort_if_no_config_available()
